requirements.txt
tornado==3.2b2
tornado-couchdb==0.2.1
facebook-sdk==0.4.0
